% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_maker.R
\name{add_random_noise}
\alias{add_random_noise}
\title{Add random noise (normally distributed) to each intensity element}
\usage{
add_random_noise(dt, intensity_cols, noise_sd)
}
\arguments{
\item{dt}{Input data frame (data.table)}

\item{intensity_cols}{Intensity column numbers}

\item{noise_sd}{Normal distribution standard deviation parameter}
}
\value{
Noise added data frame (data.table)
}
\description{
Add random noise (normally distributed) to each intensity element
}

