% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{append_original_record_idx}
\alias{append_original_record_idx}
\title{Append original record index (number, order)}
\usage{
append_original_record_idx(dt)
}
\arguments{
\item{dt}{Input data frame (data.table)}
}
\value{
Record index appended data frame (data.table)
}
\description{
Append original record index (number, order)
}

