% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pairwise_method.R
\name{apply_pairwise_method}
\alias{apply_pairwise_method}
\title{Apply pairwise (naive) method to handle dirty data such as duplicate record
and missing value problem}
\usage{
apply_pairwise_method(dt, key_cols, key_thrsh, intensity_cols, conformity_thrsh)
}
\arguments{
\item{dt}{Input data frame (data.table)}

\item{key_cols}{Key (m/z, RT ...) column numbers}

\item{key_thrsh}{Key (m/z, RT ...) column threshold with same order key_cols}

\item{intensity_cols}{Intensity column numbers}

\item{conformity_thrsh}{Conformity threshold}
}
\value{
Dirty data-free data frame (data.table)
}
\description{
Apply pairwise (naive) method to handle dirty data such as duplicate record
and missing value problem
}

