% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_processing_module.R
\name{arrange_data_using_key}
\alias{arrange_data_using_key}
\title{Rearrange data frame before processing MVP}
\usage{
arrange_data_using_key(dt, key_cols)
}
\arguments{
\item{dt}{Data frame (data.table) that want to rearrange using keys}

\item{key_cols}{Key (m/z, retention time) column numbers}
}
\value{
Rearranged data frame (data.table)
}
\description{
Rearrange data frame before processing MVP
}

