% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_processing_module.R
\name{check_condition}
\alias{check_condition}
\title{Check condition is satisfied or not.
If value difference is less than threshold than condition is satisfied}
\usage{
check_condition(current_index, candidate_index, key_col, key_thrsh)
}
\arguments{
\item{current_index}{Current record index}

\item{candidate_index}{Candidate record index}

\item{key_col}{Key (m/z, RT ...) column number}

\item{key_thrsh}{Key (m/z, RT ...) column threshold}
}
\value{
If condition satisfied thant return TRUE, else return FALSE
}
\description{
Check condition is satisfied or not.
If value difference is less than threshold than condition is satisfied
}

