% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_processing_module.R
\name{compute_intensity_similarity}
\alias{compute_intensity_similarity}
\title{Compute pairwise similarity of two records (current and candidate)}
\usage{
compute_intensity_similarity(current_index, candidate_index, intensity_cols)
}
\arguments{
\item{current_index}{Current record index}

\item{candidate_index}{Candidate record index}

\item{intensity_cols}{Intensity column numbers}
}
\value{
Similarity value (double)
}
\description{
Compute pairwise similarity of two records (current and candidate)
}

