% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_maker.R
\name{construct_simulation_data}
\alias{construct_simulation_data}
\title{Make simulation data using various parameter}
\usage{
construct_simulation_data(defect_free_dt, intensity_cols,
  maximum_duplicate = 20, noise_sd = 4, missing_rates = seq(0, 0.9, by =
  0.1))
}
\arguments{
\item{defect_free_dt}{Defect free data frame (data.table)}

\item{intensity_cols}{Intensity column numbers}

\item{maximum_duplicate}{Maximum possible duplicate number}

\item{noise_sd}{Normal distribution standard deviation parameter}

\item{missing_rates}{Missing rates sequence. e.g) c(0.1, 0.2, 0.3)}
}
\value{
Simulation data frame (data.table)
}
\description{
Make simulation data using various parameter
}

