% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{convert_input_to_na}
\alias{convert_input_to_na}
\title{Convert user specified input values to NA}
\usage{
convert_input_to_na(dt, na_expressions)
}
\arguments{
\item{dt}{Input data frame (data.table)}

\item{na_expressions}{The values that user want to convert to NA}
}
\value{
NA mapped data frame (data.table)
}
\description{
Convert user specified input values to NA
}

