% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_maker.R
\name{extract_defect_free_table}
\alias{extract_defect_free_table}
\title{Extract all records don't have any missing value (Defect free records)}
\usage{
extract_defect_free_table(dt, intensity_cols)
}
\arguments{
\item{dt}{Input data frame (data.table)}

\item{intensity_cols}{Intensity column numbers}
}
\value{
Data frame (data.table) consist of defect free records
}
\description{
Extract all records don't have any missing value (Defect free records)
}

