% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_processing_module.R
\name{extract_duplicate_candidate}
\alias{extract_duplicate_candidate}
\title{Extract duplicate candidates related current record (using index number)}
\usage{
extract_duplicate_candidate(current_index, key_cols, key_thrshs)
}
\arguments{
\item{current_index}{Current record index}

\item{key_cols}{Key (m/z, RT can be both or only one) column numbers}

\item{key_thrshs}{Key (m/z, RT same order of key_cols) column thresholds}
}
\value{
Duplicate candidates related to current record
}
\description{
Extract duplicate candidates related current record (using index number)
}

