% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{find_non_intensity_cols}
\alias{find_non_intensity_cols}
\title{Find non-intensity columns in data frame}
\usage{
find_non_intensity_cols(whole_cols, intensity_cols)
}
\arguments{
\item{whole_cols}{Whole column numbers}

\item{intensity_cols}{Intensity column numbers}
}
\value{
Non intensity column numbers
}
\description{
Find non-intensity columns in data frame
}

