% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_processing_module.R
\name{inspect_conditions}
\alias{inspect_conditions}
\title{Inspect current record and candidate record are satisfied with conditions}
\usage{
inspect_conditions(current_index, candidate_index, key_cols, key_thrshs)
}
\arguments{
\item{current_index}{Current record index}

\item{candidate_index}{Candidate record index}

\item{key_cols}{Key (m/z, RT) column numbers}

\item{key_thrshs}{Key (m/z, RT) column thresholds}
}
\value{
Condition statement string
}
\description{
Inspect current record and candidate record are satisfied with conditions
}

