\name{letters}
\alias{letters}
\alias{lettersymbols}
\alias{a}
\alias{b}
\alias{c}
\alias{d}
\alias{e}
\alias{f}
\alias{g}
\alias{h}
\alias{i}
\alias{j}
\alias{k}
\alias{l}
\alias{m}
\alias{n}
\alias{o}
\alias{p}
\alias{q}
\alias{r}
\alias{s}
\alias{t}
\alias{u}
\alias{v}
\alias{w}
\alias{x}
\alias{y}
\alias{z}
\alias{A}
\alias{B}
\alias{C}
\alias{D}
\alias{E}
\alias{F}
\alias{G}
\alias{H}
\alias{I}
\alias{J}
\alias{K}
\alias{L}
\alias{M}
\alias{N}
\alias{O}
\alias{P}
\alias{Q}
\alias{R}
\alias{S}
\alias{T}
\alias{U}
\alias{V}
\alias{W}
\alias{X}
\alias{Y}
\alias{Z}
\docType{data}
\title{Single-letter symbols}
\description{
  Variables \code{a}, \code{b},\ldots \code{z} are given their \code{mvp}
  semantic meaning.
  }
\usage{data(lettersymbols)}
\details{

  Twenty-six variables a-z and twenty-six variables A-Z are defined as
  their \code{mvp} semantic equivalent:
  
  \preformatted{
    a <- as.mvp("a")
    ...
    z <- as.mvp("z")
    A <- as.mvp("A")
    ...
    Z <- as.mvp("Z")
  }

  These objects can be generated by running script \code{inst/symb.Rmd},
  which includes some further discussion and technical documentation and
  creates file \code{lettersymbols.rda} which resides in the
  \code{data/} directory.

  Letters \code{c}, \code{q}, and \code{t}, \code{F}, and \code{T} might
  pose difficulties.

}
\author{Robin K. S. Hankin}
\examples{

data(lettersymbols)
(a+b)*(a-b)

(x + y + z)^3 - 3*(x + y + z)*(x*y + x*z + y*z) + 3*x*y*z

((x+y)^2 + y^2)*((x-y)^2 + y^2) # Sophie Germain's identity (so-called)

}
\keyword{datasets}

