% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_maker.R
\name{make_random_missing_value}
\alias{make_random_missing_value}
\title{Make random missing value to intensity element}
\usage{
make_random_missing_value(dt, intensity_cols, missing_rates)
}
\arguments{
\item{dt}{Input data frame (data.table)}

\item{intensity_cols}{Intensity column numbers}

\item{missing_rates}{Missing rates sequence. e.g) c(0.1, 0.2, 0.3)}
}
\value{
Randomly missing value made data frame
}
\description{
Make random missing value to intensity element
}

