% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_maker.R
\name{make_record_duplicated_table}
\alias{make_record_duplicated_table}
\title{To simulate real world data, make record duplicated data frame (data.table)
To test various situation, user can adjust maximum duplicate number}
\usage{
make_record_duplicated_table(dt, maximum_duplicate)
}
\arguments{
\item{dt}{Input data frame (data.table)}

\item{maximum_duplicate}{Maximum possible duplicate number}
}
\value{
Record duplicated data frame
}
\description{
To simulate real world data, make record duplicated data frame (data.table)
To test various situation, user can adjust maximum duplicate number
}

