% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_maker.R
\name{manipulate_table_index}
\alias{manipulate_table_index}
\title{Before testing simulation restoration rate, we need to manipulated record
index. manipuliate_table_index can handle this procedure.}
\usage{
manipulate_table_index(record, current_number, maximum_number)
}
\arguments{
\item{record}{Current record}

\item{current_number}{Current duplicate number (1 to maximum_number)}

\item{maximum_number}{Maximum duplicate number}
}
\value{
Table index manipulated record
}
\description{
Before testing simulation restoration rate, we need to manipulated record
index. manipuliate_table_index can handle this procedure.
}

