% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pairwise_method.R
\name{merge_pairwise_record}
\alias{merge_pairwise_record}
\title{Merge pairwise records (two records), manipulate non-intensity part and
intensity part}
\usage{
merge_pairwise_record(current_index, matching_index, intensity_cols)
}
\arguments{
\item{current_index}{Current record index}

\item{matching_index}{Best similar record index}

\item{intensity_cols}{Intensity column numbers}
}
\description{
Merge pairwise records (two records), manipulate non-intensity part and
intensity part
}

