% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_clique_method.R
\name{merge_records_of_clique}
\alias{merge_records_of_clique}
\title{Title}
\usage{
merge_records_of_clique(record_indexes_of_clique, intensity_cols)
}
\arguments{
\item{record_indexes_of_clique}{Integer record indexes of clique
e.g. c(3123, 3127, 3130, 3133)}

\item{intensity_cols}{Intensity columns}
}
\value{
Final index which should be selected
}
\description{
Title
}

