% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{preprocess_input_data}
\alias{preprocess_input_data}
\title{Before applying MVP algorithm, preprocess input data such as na converting}
\usage{
preprocess_input_data(input, intensity_cols, na_expressions)
}
\arguments{
\item{input}{Input file path that want to process MVP}

\item{intensity_cols}{Intensity column numbers}

\item{na_expressions}{The values that user want to convert to NA}
}
\value{
Preprocessed data frame (data.table) that can use MVP algorithm
}
\description{
Before applying MVP algorithm, preprocess input data such as na converting
}

