% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pairwise_method.R
\name{process_pairwise_merge}
\alias{process_pairwise_merge}
\title{Pairwise merge 1 iteration and return after 1 iteration result}
\usage{
process_pairwise_merge(dt, key_cols, key_thrshs, intensity_cols,
  conformity_thrsh)
}
\arguments{
\item{dt}{Input data frame (data.table)}

\item{key_cols}{Key (m/z, RT ...) column numbers}

\item{key_thrshs}{Key (m/z, RT ...) column thresholds}

\item{intensity_cols}{Intensity column numbers}

\item{conformity_thrsh}{conformity threshold}
}
\value{
After one iteration result
}
\description{
Pairwise merge 1 iteration and return after 1 iteration result
}

